Revert "update: move contexts and providers to separate folders"
This reverts commit cba24c10df9945f1e16af4ee4c7561ed0d94ed9c.
