--- src/context/providers/MainContextProvider/index.ts
@@ -1,1 +1,0 @@
-export * from './MainContextProvider'
