--- src/context/providers/index.ts
@@ -1,2 +1,0 @@
-export * from './MainContextProvider'
-export * from './SkillsContextProvider'
